backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
motor>=3.6.0
pymongo>=4.9
redis==5.0.1
qdrant-client==1.7.1
httpx==0.26.0
sentence-transformers>=2.3.0
openai==1.10.0
structlog==24.1.0
prometheus-client==0.19.0
python-multipart==0.0.6
pytest==7.4.4
playwright==1.41.0
beautifulsoup4==4.12.3
rq==1.16.0
duckduckgo-search>=4.0.0
tavily-python>=0.5.0
feedparser>=6.0.0
